#843 fix warnings for 64 bit

+++ Paco-iOS/Paco/source/core/scheduling/PacoScheduleGenerator+Daily.h
@@ -18,7 +18,7 @@
 @interface PacoScheduleGenerator (Daily)
 
 + (NSArray*)nextDatesForDailyExperiment:(PacoExperiment*)experiment
-                             numOfDates:(NSInteger)numOfDates
+                             numOfDates:(int)numOfDates
                                fromDate:(NSDate*)fromDate;
 
 @end
